requirements.txt
streamlit
sqlalchemy
snowflake-connector-python
streamlit_card